aider: Fix the `EpubHtml` object not being iterable error by setting `book.toc` as a list of tuples in the `create_epub` function in `extract.py`.
# Aider chat conversation:

USER: Traceback (most recent call last):
  File "C:\GitRoot\PaperAgent\extract.py", line 172, in <module>
    create_epub(papers, f"{DATE}.epub")
  File "C:\GitRoot\PaperAgent\extract.py", line 154, in create_epub
    epub.write_epub(filename, book, {})
  File "C:\GitRoot\PaperAgent\venv\lib\site-packages\ebooklib\epub.py", line 1746, in write_epub
    epub.write()
  File "C:\GitRoot\PaperAgent\venv\lib\site-packages\ebooklib\epub.py", line 1369, in write
    self._write_items()
  File "C:\GitRoot\PaperAgent\venv\lib\site-packages\ebooklib\epub.py", line 1354, in _write_items
    self.out.writestr('%s/%s' % (self.book.FOLDER_NAME, item.file_name), self._get_ncx())
  File "C:\GitRoot\PaperAgent\venv\lib\site-packages\ebooklib\epub.py", line 1345, in _get_ncx
    _create_section(nav_map, self.book.toc, 0)
  File "C:\GitRoot\PaperAgent\venv\lib\site-packages\ebooklib\epub.py", line 1305, in _create_section
    uid = _create_section(np, subsection, uid + 1)
  File "C:\GitRoot\PaperAgent\venv\lib\site-packages\ebooklib\epub.py", line 1279, in _create_section
    for item in items:
TypeError: 'EpubHtml' object is not iterable

ASSISTANT: The error message indicates that the `EpubHtml` object is not iterable. This is because the `book.toc` is expected to be a list of `EpubHtml` objects or a list of tuples where each tuple represents a section and contains two elements: the section title (an `EpubHtml` object) and a list of subsections (a list of `EpubHtml` objects). In the current code, `book.toc` is set as a tuple of an `EpubHtml` object and a list of `EpubHtml` objects, which is causing the error.

To fix this, we need to modify the `create_epub` function in `extract.py` to set `book.toc` as a list of tuples where each tuple represents a section and contains two elements: the section title (an `EpubHtml` object) and a list of subsections (a list of `EpubHtml` objects).

Here is the *SEARCH/REPLACE* block:

extract.py
```python
<<<<<<< SEARCH
    # Create the TOC
    book.toc = (date_chapter, paper_chapters)
=======
    # Create the TOC
    book.toc = [(date_chapter, paper_chapters)]
>>>>>>> REPLACE
```