aider: Fix extract_paper_details function to return a Paper object directly.
# Aider chat conversation:

USER: ((venv) ) REDMOND.yuwan@GCRSANDBOX374:~/GitRoot/PaperAgent$ python download.py
Downloading data for 2024-01-01
Traceback (most recent call last):
  File "/home/yuwan/GitRoot/PaperAgent/download.py", line 22, in <module>
    download_data(start_date, end_date)
  File "/home/yuwan/GitRoot/PaperAgent/download.py", line 12, in download_data
    get_papers(date_str)
  File "/home/yuwan/GitRoot/PaperAgent/paper.py", line 139, in get_papers
    papers = extract_papers(date)
  File "/home/yuwan/GitRoot/PaperAgent/paper.py", line 118, in extract_papers
    paper_obj = extract_paper_details(paper_div, date)
  File "/home/yuwan/GitRoot/PaperAgent/paper.py", line 51, in extract_paper_details
    return {
TypeError: unhashable type: 'Paper'

ASSISTANT: The error message indicates that there is a problem with the `extract_paper_details` function in the `paper.py` file. It seems that the function is returning a set with a `Paper` object inside, which is not hashable and thus causing a `TypeError`. The correct behavior should be to return a `Paper` object directly, not wrapped in a set or any other collection. Let's fix this issue by updating the `extract_paper_details` function.

Here is the diff for the required changes:

```diff
--- paper.py
+++ paper.py
@@ ... @@
 def extract_paper_details(paper_div, date):
     title = paper_div.find('h2', class_='title').find('a', class_='title-link').get_text(strip=True)
     link = paper_div.find('h2', class_='title').find('a', href=True)['href']
     authors = ' ; '.join([author.get_text(strip=True) for author in paper_div.find('p', class_='authors').find_all('span', class_='author')])
     abstract = paper_div.find('p', class_='summary').get_text(strip=True)
-    return {
+    return Paper(date=date, title=title, link=link, authors=authors, abstract=abstract)
```

This change will ensure that the `extract_paper_details` function returns a `Paper` object as intended.